Fixed drawing of images to expand to the full combatant size.
I had to add a mask that is shaped like the combatant, but the way svg
works is a little odd. I tried a bunch of ways to make it work, and this
is the only way I've been able to get it to actually work. I tried
combatants of various sizes, as well as ones without tokens, and it all
works fine.